Closes #3 infix fmap operator.
This is a non-backwards compatible change. While having a unit shortcut is cool because adding a binop on a class is cool, it generally complicated things and made the code confusing. Also the bitwise and has a lower precedence than the multiplication operator. However, modulo has the same precedence as mul, meaning Python will parse the expression left to right with no unneeded parens, bringing pynads closer Haskell in terms of syntax.