Rename kwargs_decorator to optional_kwargs. Added descriptor based implementation of optional_kwargs.
optional_kwargs fits the intent of the decorator better. method_optional_kwargs is a descriptor based implementation of it for use with instance methods, classmethods and staticmethods (optional_kwargs also works with staticmethod, but under it rather than above it). method_optional_kwargs doesn't work with other sorts of descriptors yet, but can be used to wrap classmethod and staticmethod on the top level.